It now checks to make sure you won't teleport into a wall
It now checks to make sure you won't teleport into a wall, so that you
won't ever get stuck in the walls again. However, it's very annoying at
the moment because it doesn't have any forgiveness for misplaced clicks
yet. So for now, I'm putting this on the back burner.