Dragging camera off Merky is no longer allowed
When the camera is dragged around, it is limited to the areas that show
Merky on screen.

NOTE: It is possible to move the camera such that Merky is on the edge
of the screen, and zoom in, causing him to not be on screen.

NOTE: How fast the camera moves effects how far you can move the camera
to a small degree. It's not a big issue but should be taken care of