Cracked ground explodes into rubble
When a cracked ground is destroyed, it now spawns rubble pieces in its
place.
Rubble pieces that spawn overlapping a static collidable object become
static.
Rubble pieces can be moved around afterwards, and can trigger weight
switches currently.

NOTE: Some of the level will have to be redesigned to accomodate for the
rubble pieces. One puzzle in particular can be easily circumvented with
the rubble pieces.