Finally got it working: CrackedGround gets saved!
The issue was that the subtype was not saving out the same number of
things as the supertype. As far as I understand, It would save out as
the subtype but read in as the supertype. Once I got the subtype writing
the same stuff that the supertype was reading, it all worked perfectly!