Patched hidden areas not going away upon opening
There's a bug where the hidden areas you've discovered last session
still appear and block your way when you close and then open another
session.

It's been fixed, but is not optimal. Optimally, the memory would be on
the Hidden Area parent and not on the things that trigger it to be seen.