Fixed #94 Shield Bubble TRI size [QUICK FIX]
As with the previous commit, this commit relies on a pre-calculated
value instead of figuring out the correct values upon start up. This is
due to some strangeness in the timing of the Start() method call, along
with some other things that are unknown about how Unity works in regards
to sizes.

It fixes the issue, altho if the sprite asset for the
TeleportRangeIndicator ever changes size, these values will need to be
updated.